Moved the files to their respective directories
