+++ main-src/src/client/App.tsx
@@ -1,9 +1,9 @@
 import "./App.css";
 import { BrowserRouter as Router, Routes, Route} from "react-router-dom";
-import Home from "./homeScreen";
-import Login from "./Login"; 
-import Signup from "./signUpScreen";
-import Forgotpass from "./forgotPassScreen";
+import Home from "./components/homeScreen";
+import Login from "./components/Login"; 
+import Signup from "./components/signUpScreen";
+import Forgotpass from "./components/forgotPassScreen";
 import { Header } from "./components/Header";
 import { Main } from "./components/Main";
 import { Footer } from "./components/Footer";
@@ -25,9 +25,9 @@ function App() {
           <Route path="/signup" element={<Signup />} />
           <Route path="/home" element={<Home />} />
           <Route path="/forgotpass" element={<Forgotpass />} />
-          <Route path="/scheduleworkout" element={<Scheduleworkout />} /> 
+          {/* <Route path="/scheduleworkout" element={<Scheduleworkout />} /> 
           <Route path="/contactus" element={<Contactus />} />
-          <Route path="/aboutus" element={<Aboutus />} />
+          <Route path="/aboutus" element={<Aboutus />} /> */}
           <Route path="/addMealEntry" element={<AddMealEntry />} />
           <Route path="*" element={<NotFoundPage />} />
           <Route path="/calendar" element={<CalendarPage />} />
@@ -35,6 +35,4 @@ function App() {
       </div>
   ); 
 }
-export default App;
-
 export default App;
