+++ main-src/src/client/components/forgotPassScreen.tsx
@@ -1,4 +1,4 @@
-import App from "./App"; 
+import App from "../App"; 
 import "./components/forgotPass.css"
 
 export const Forgotpass = () => {
